Move remaining types to check engine
CMK-13768
