+++ cmk/base/agent_based/checking/_checking.py
@@ -23,7 +23,6 @@ from cmk.utils.resulttype import Result
 from cmk.utils.sectionname import SectionName
 from cmk.utils.structured_data import TreeStore
 from cmk.utils.timeperiod import check_timeperiod, timeperiod_active, TimeperiodName
-from cmk.utils.type_defs import ServiceName
 
 from cmk.snmplib.type_defs import AgentRawData, SNMPRawData
 
@@ -42,7 +41,7 @@ from cmk.checkengine import (
     SummarizerFunction,
 )
 from cmk.checkengine.check_table import ConfiguredService
-from cmk.checkengine.checking import CheckPluginName
+from cmk.checkengine.checking import CheckPluginName, ServiceName
 from cmk.checkengine.checkresults import (
     ActiveCheckResult,
     MetricTuple,
