+++ cmk/base/agent_based/discovery/preview.py
@@ -14,7 +14,6 @@ from cmk.utils.labels import DiscoveredHostLabelsStore, HostLabel, ServiceLabel
 from cmk.utils.log import console
 from cmk.utils.sectionname import SectionName
 from cmk.utils.timeperiod import timeperiod_active
-from cmk.utils.type_defs import Item, ServiceName
 
 from cmk.automations.results import CheckPreviewEntry
 
@@ -28,7 +27,7 @@ from cmk.checkengine import (
     SummarizerFunction,
 )
 from cmk.checkengine.check_table import ConfiguredService
-from cmk.checkengine.checking import CheckPluginName
+from cmk.checkengine.checking import CheckPluginName, Item, ServiceName
 from cmk.checkengine.checkresults import ActiveCheckResult, ServiceCheckResult
 from cmk.checkengine.discovery import (
     analyse_cluster_labels,
