+++ cmk/base/api/agent_based/value_store/_utils.py
@@ -23,10 +23,9 @@ import cmk.utils.store as store
 from cmk.utils.exceptions import MKGeneralException
 from cmk.utils.hostaddress import HostName
 from cmk.utils.log import logger
-from cmk.utils.type_defs import Item
 
 from cmk.checkengine.check_table import ServiceID
-from cmk.checkengine.checking import CheckPluginName
+from cmk.checkengine.checking import CheckPluginName, Item
 
 _PluginName = str
 _UserKey = str
