+++ cmk/base/automations/check_mk.py
@@ -61,7 +61,6 @@ from cmk.utils.paths import (
 )
 from cmk.utils.timeout import Timeout
 from cmk.utils.timeperiod import timeperiod_active
-from cmk.utils.type_defs import CheckPluginNameStr
 
 from cmk.automations.results import (
     ActiveCheckResult,
@@ -110,7 +109,7 @@ from cmk.fetchers.filecache import FileCacheOptions
 from cmk.fetchers.snmp import make_backend as make_snmp_backend
 
 from cmk.checkengine import parse_raw_data, plugin_contexts, SourceType
-from cmk.checkengine.checking import CheckPluginName
+from cmk.checkengine.checking import CheckPluginName, CheckPluginNameStr
 from cmk.checkengine.discovery import AutocheckEntry, AutocheckServiceWithNodes, DiscoveryMode
 from cmk.checkengine.discovery import DiscoveryResult as SingleHostDiscoveryResult
 from cmk.checkengine.submitters import ServiceDetails, ServiceState
