+++ cmk/base/core_config.py
@@ -26,10 +26,9 @@ from cmk.utils.licensing.handler import LicenseState, LicensingHandler
 from cmk.utils.licensing.helper import get_licensed_state_file_path, write_licensed_state
 from cmk.utils.paths import core_helper_config_dir
 from cmk.utils.store import load_object_from_file, lock_checkmk_configuration, save_object_to_file
-from cmk.utils.type_defs import Item, ServiceName
 
 from cmk.checkengine.check_table import ConfiguredService, ServiceID
-from cmk.checkengine.checking import CheckPluginName
+from cmk.checkengine.checking import CheckPluginName, Item, ServiceName
 from cmk.checkengine.parameters import TimespecificParameters
 
 import cmk.base.api.agent_based.register as agent_based_register
