+++ cmk/base/export.py
@@ -12,9 +12,8 @@ from the configuration.
 from cmk.utils.hostaddress import HostName
 from cmk.utils.labels import Labels
 from cmk.utils.rulesets.ruleset_matcher import RulesetMatcher, RulesetMatchObject
-from cmk.utils.type_defs import CheckPluginNameStr, Item, ServiceName
 
-from cmk.checkengine.checking import CheckPluginName
+from cmk.checkengine.checking import CheckPluginName, CheckPluginNameStr, Item, ServiceName
 
 import cmk.base.config as config
