+++ cmk/bi/aggregation.py
@@ -10,7 +10,7 @@ from typing import Any
 from cmk.utils.hostaddress import HostName
 
 # TODO: fix duplicate type def. the original type def is in gui-managed (module layer violation)
-from cmk.utils.type_defs import ServiceName
+from cmk.checkengine.checking import ServiceName
 
 from cmk.bi.lib import (
     ABCBICompiledNode,
