+++ cmk/checkengine/discovery/filters.py
@@ -7,7 +7,8 @@ from collections.abc import Callable
 from typing import Any, NamedTuple
 
 from cmk.utils.regex import regex
-from cmk.utils.type_defs import ServiceName
+
+from cmk.checkengine.checking import ServiceName
 
 ServiceFilter = Callable[[ServiceName], bool]
