+++ cmk/checkengine/error_handling.py
@@ -17,11 +17,11 @@ from cmk.utils.exceptions import (
     MKTimeout,
 )
 from cmk.utils.hostaddress import HostName
-from cmk.utils.type_defs import CheckPluginNameStr, ServiceName
 
 from cmk.snmplib.type_defs import SNMPBackendEnum
 
 from cmk.checkengine import crash_reporting
+from cmk.checkengine.checking import CheckPluginNameStr, ServiceName
 
 from .submitters import ServiceState
