+++ cmk/checkengine/submitters.py
@@ -19,8 +19,8 @@ from cmk.utils.exceptions import MKGeneralException
 from cmk.utils.hostaddress import HostName
 from cmk.utils.log import console
 from cmk.utils.timeout import Timeout
-from cmk.utils.type_defs import ServiceName
 
+from cmk.checkengine.checking import ServiceName
 from cmk.checkengine.checkresults import ServiceCheckResult
 
 _CacheInfo = tuple[int, int]
