+++ cmk/gui/dashboard/dashlet/dashlets/graph.py
@@ -13,7 +13,8 @@ import livestatus
 from cmk.utils.exceptions import MKGeneralException
 from cmk.utils.hostaddress import HostName
 from cmk.utils.macros import MacroMapping
-from cmk.utils.type_defs import MetricName
+
+from cmk.checkengine.checking import MetricName
 
 import cmk.gui.sites as sites
 from cmk.gui.exceptions import MKMissingDataError, MKUserError
