+++ cmk/gui/plugins/metrics/rrd_fetch.py
@@ -16,7 +16,8 @@ import cmk.utils.version as cmk_version
 from cmk.utils.exceptions import MKGeneralException
 from cmk.utils.hostaddress import HostName
 from cmk.utils.prediction import livestatus_lql, TimeSeries, TimeSeriesValues
-from cmk.utils.type_defs import ServiceName
+
+from cmk.checkengine.checking import ServiceName
 
 import cmk.gui.plugins.metrics.timeseries as ts
 import cmk.gui.sites as sites
