+++ cmk/gui/type_defs.py
@@ -31,8 +31,9 @@ from cmk.utils.labels import Labels
 from cmk.utils.notify_types import DisabledNotificationsOptions, EventRule
 from cmk.utils.store.host_storage import ContactgroupName
 from cmk.utils.structured_data import SDPath
-from cmk.utils.type_defs import MetricName, ServiceName
 from cmk.utils.user import UserId
+
+from cmk.checkengine.checking import MetricName, ServiceName
 
 from cmk.gui.exceptions import FinalizeRequest
 from cmk.gui.utils.speaklater import LazyString
