+++ cmk/gui/wato/pages/check_catalog.py
@@ -15,7 +15,8 @@ from typing import Any, overload
 
 import cmk.utils.man_pages as man_pages
 from cmk.utils.man_pages import ManPageCatalogPath
-from cmk.utils.type_defs import CheckPluginNameStr
+
+from cmk.checkengine.checking import CheckPluginNameStr
 
 from cmk.gui.breadcrumb import Breadcrumb, BreadcrumbItem
 from cmk.gui.exceptions import MKUserError
