+++ cmk/gui/wato/pages/pattern_editor.py
@@ -9,7 +9,8 @@ from collections.abc import Collection, Iterable
 
 from cmk.utils.hostaddress import HostName
 from cmk.utils.labels import Labels
-from cmk.utils.type_defs import CheckPluginNameStr, Item, ServiceName
+
+from cmk.checkengine.checking import CheckPluginNameStr, Item, ServiceName
 
 # Tolerate this for 1.6. Should be cleaned up in future versions,
 # e.g. by trying to move the common code to a common place
