+++ cmk/gui/watolib/check_mk_automations.py
@@ -12,10 +12,11 @@ from cmk.utils.diagnostics import DiagnosticsCLParameters
 from cmk.utils.exceptions import MKGeneralException
 from cmk.utils.hostaddress import HostName
 from cmk.utils.labels import HostLabel
-from cmk.utils.type_defs import ServiceName
 
 from cmk.automations import results
 from cmk.automations.results import SetAutochecksTable
+
+from cmk.checkengine.checking import ServiceName
 
 from cmk.gui.i18n import _
 from cmk.gui.site_config import site_is_local
