+++ tests/unit/cmk/gui/plugins/metrics/test_metrics_utils.py
@@ -10,7 +10,8 @@ import pytest
 from tests.unit.cmk.gui.conftest import SetConfig
 
 from cmk.utils.exceptions import MKGeneralException
-from cmk.utils.type_defs import MetricName
+
+from cmk.checkengine.checking import MetricName
 
 import cmk.gui.metrics as metrics
 from cmk.gui.config import active_config
